dnsdist: Move internal DoQ structures to doq.cc
